develop panels for plots
Don't plot years without data
Create multiple panels if the number of plots
per panel is greater than the maximum provided.
The new default is to create 2 x 2 panels, so if you
have more than four years with data, two or more
files will be saved.
The default behaviour before generated at least 16 panels
per file and if more than one file was created it was saved
with the same name as the previous, so you just ended up with
the terminal years.