Create PESC_example_red_grouper.rda
I created an "example" object so that we can demonstrate the estimation of predator-expanded-stomach-contents (PESCs) and diet proportions to VAST users, using West Florida Shelf red grouper as an example. The "example" object contains information about the "sampling_data", "Region" and "strata.limits", but also the "input_grid". This is because the option  Region="Other" turned out to be buggy for West Florida Shelf red grouper, so we are instead defining Region="User" and providing the West Florida Shelf grid to VAST users via the "example" object.